fix: reactive value being tree shaked accidentally
