+++ .changeset/bright-mammals-judge.md
@@ -1,0 +1,5 @@
+---
+'@lynx-js/lynx-ui-common': patch
+---
+
+Fix an issue where using ReactiveValue causes an MTS error
