+++ packages/lynx-ui-common/src/reactive/index.ts
@@ -2,9 +2,6 @@
 // Licensed under the Apache License Version 2.0 that can be found in the
 // LICENSE file in the root directory of this source tree.
 
-// Core reactive value system
-// export { createMainThreadReactiveValue } from './reactiveValue'
-
 // Not Ready yet
 // export { derived } from './derived'
